moved actions to separate hook
